requirements.txt
Django
djangorestframework
gunicorn
mysqlclient
django-push-notifications
django-crontab
psycopg2
requests
beautifulsoup4
